Use only preconfigured audio effects for Syberia
- Set preconfigured Audio Effects and some lib backup to syberia os only

-  Do not remove DIRAC from xenonhd, cause xenonhd still not implemented this function. If any other ROM still not then i'll add it as exception (if i found it :p)

- Update TAG build to v0.4

+++ dirac.sh
@@ -2,12 +2,11 @@
 #
 
 #Mounting Your System
-mounting(){
 	mount -o remount,rw /system
-}
-
+	
 #Backup your stock libs file
-backup(){
+if [ "`grep ro.product.name=syberia_mido /system/build.prop`" ];
+then 
 	mkdir /sdcard/backup
 	mkdir /sdcard/backup/lib
 	mkdir /sdcard/backup/lib64
@@ -33,17 +32,53 @@ backup(){
 	cp /vendor/etc/audio_effects.xml /sdcard/backup/vendor/etc/audio_effects.xml
 	cp /vendor/etc/audio_output_policy.conf /sdcard/backup/vendor/etc/audio_output_policy.conf
 	cp /vendor/etc/audio/audio_policy.conf /sdcard/backup/vendor/etc/audio/audio_policy.conf
-}
+else
+	mkdir /sdcard/backup
+	mkdir /sdcard/backup/lib
+	mkdir /sdcard/backup/lib64
+	mkdir /sdcard/backup/vendor
+	mkdir /sdcard/backup/vendor/etc
+	cp /system/lib/libaudioclient.so /sdcard/backup/lib/libaudioclient.so
+	cp /system/lib/libaudioeffect_jni.so /sdcard/backup/lib/libaudioeffect_jni.so
+	cp /system/lib/libaudioflinger.so /sdcard/backup/lib/libaudioflinger.so
+	cp /system/lib/libaudiopolicyenginedefault.so /sdcard/backup/lib/libaudiopolicyenginedefault.so
+	cp /system/lib/libaudiopolicymanager.so /sdcard/backup/lib/libaudiopolicymanager.so
+	cp /system/lib/libaudiopolicymanagerdefault.so /sdcard/backup/lib/libaudiopolicymanagerdefault.so 
+	cp /system/lib/libaudiopolicyservice.so /sdcard/backup/lib/libaudiopolicyservice.so
+	cp /system/lib/libaudioprocessing.so /sdcard/backup/lib/libaudioprocessing.so
+	cp /system/lib/libsoundpool.so /sdcard/backup/lib/libsoundpool.so
+	cp /system/lib/libsoundtrigger.so /sdcard/backup/lib/libsoundtrigger.so
+	cp /system/lib/libsoundtriggerservice.so /sdcard/backup/lib/libsoundtriggerservice.so
+	cp /system/lib64/libaudioclient.so /sdcard/backup/lib64/libaudioclient.so
+	cp /system/lib64/libaudiopolicyenginedefault.so /sdcard/backup/lib64/libaudiopolicyenginedefault.so
+	cp /system/lib64/libaudiopolicymanager.so /sdcard/backup/lib64/libaudiopolicymanager.so
+	cp /system/lib64/libaudiopolicymanagerdefault.so /sdcard/backup/lib64/libaudiopolicymanagerdefault.so
+	cp /system/lib64/libaudiopolicyservice.so /sdcard/backup/lib64/libaudiopolicyservice.so
+	cp /system/lib64/libaudioroute.so /sdcard/backup/lib64/libaudioroute.so 
+fi
 
 #Remove DIRAC Audio Effetcs From System
-dirac(){
-rm /system/lib/soundfx/libdirac.so
-rm /system/lib/libDiracAPI_SHARED.so
-rm /system/vendor/etc/diracmobile.config
-sed -i '/persist.audio.dirac.speaker=true/d' /vendor/build.prop
-}
+#XenonHD ROM not integrate DIRAC, so not remove it
+if [ "`grep ro.product.name=xenonhd_mido /system/build.prop`" ];
+then
+	echo "Do not remove dirac from xenonhd"
+else
+	rm /system/lib/soundfx/libdirac.so
+	rm /system/lib/libDiracAPI_SHARED.so
+	rm /system/vendor/etc/diracmobile.config
+	sed -i '/persist.audio.dirac.speaker=true/d' /vendor/build.prop
+fi
 
-#Execute Main Program
-mounting
-backup
-dirac
+#Copy modified audio_effects.xml for syberia os only 
+#Syberia have updated this config but other ROMs not, so copy only for preconfigured syberia only
+#Will check other ROM later.
+if [ "`grep ro.product.name=syberia_mido /system/build.prop`" ];
+then 
+    rm /system/vendor/etc/audio_effects.xml
+	mv /system/vendor/etc/syberia_audio_effects.xml /system/vendor/etc/audio_effects.xml
+	chmod 0644 /system/vendor/etc/audio_effects.xml
+else
+	rm /system/vendor/etc/audio_effects.xml
+	mv /system/vendor/etc/default_audio_effects.xml /system/vendor/etc/audio_effects.xml
+	chmod 0644 /system/vendor/etc/audio_effects.xml
+fi
